search_api/requirements.txt
fastapi
uvicorn
weaviate-client
python-dotenv
sentence-transformers
